Support representation of alternate rate structures
Create and process new data (./scout/supporting_data/sub_fed/rates.csv) to allow users to represent the effects of electrification-friendly electricity rates in the Scout forecast.

The function import_state_data, was broadened to process input data from both user incentives and rates files in one place.

Users may modify rates for a given electricity segment that the measure applies to by reducing the volumetric rate – either in absolute terms, e.g., $/kWH reduction; or in relative terms, e.g., % reduction of existing volumetric rate – which is then carried through to all subsequent calculations of measure energy costs.

Users may also specify when there are increases in fixed rates in alternate rate schemes to offset the decrease in volumetric rates; however, these increases are not currently reflected on measure costs, since a flat rate applies to the whole electricity bill and all electricity measures, and is not readily attributable to a single measure’s calculations. Future updates might modify the handling of increased fixed charges, and for now there are placeholders in the code to mark where this handling would be added.

As with other adoption drivers, a year range and applicability factor may be applied to the alternate rates.

Addresses issue #411


Former-commit-id: ce5422b9262365f9772e4b59b283f64b38f88937