Add handling of limited linked seg data prep
To limit data output, technical potential contributing mseg data preparation is restricted to only those data needed for competition in run.py when the user is executing a max adoption potential scenario only. This caused a problem with the new code to link costs across multiple segments in a measure, as the code was trying to pull from linked data fields that did not exist/were restricted. This issue has been fixed.


Former-commit-id: e168d55493899f9e2ff9f15cab71474a64218d73